# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dashboard/pools": {
            "get": {
                "description": "get Pools data of dezswap (address, tvl, volume, fee, apr)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dezswap's Pools",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dashboard.PoolRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/dashboard/recent": {
            "get": {
                "description": "get Recent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Recent 24H data with it's change rate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dashboard.RecentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/dashboard/statistics": {
            "get": {
                "description": "get Statistic data of dezswap (address count, tx count, fee)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dezswap's statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dashboard.StatisticRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/dashboard/tokens": {
            "get": {
                "description": "get Tokens data of dezswap (address, price, priceChange, volume, tvl)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dezswap's Tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dashboard.TokenRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/dashboard/tvls": {
            "get": {
                "description": "get TVLs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "TVLs of dezswap selected duration",
                "parameters": [
                    {
                        "enum": [
                            "year",
                            "quarter",
                            "month"
                        ],
                        "type": "string",
                        "description": "default(empty) value is all",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dashboard.TVLRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/dashboard/txs": {
            "get": {
                "description": "get Transactions data of dezswap (action, totalValue, asset0amount, asset1amount, time)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dezswap's Transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dashboard.TxRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/dashboard/volumes": {
            "get": {
                "description": "get Volumes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Volumes of user selected duration",
                "parameters": [
                    {
                        "enum": [
                            "year",
                            "quarter",
                            "month"
                        ],
                        "type": "string",
                        "description": "default(empty) value is all",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dashboard.VolumeRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/pairs": {
            "get": {
                "description": "get Pairs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "All Pairs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PairsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/pairs/{address}": {
            "get": {
                "description": "get Pair by Address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a pair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pair Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.PairRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.PairRes": {
            "type": "object",
            "properties": {
                "asset_decimals": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "asset_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dezswap.AssetInfoTokenRes"
                    }
                },
                "contract_addr": {
                    "type": "string"
                },
                "liquidity_token": {
                    "type": "string"
                }
            }
        },
        "controller.PairsRes": {
            "type": "object",
            "properties": {
                "pairs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.PairRes"
                    }
                }
            }
        },
...
        "dashboard.AddressCountRes": {
            "type": "object",
            "properties": {
                "addressCount": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dashboard.FeeRes": {
            "type": "object",
            "properties": {
                "fee": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "dashboard.PoolRes": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "apr": {
                    "type": "string"
                },
                "fee": {
                    "type": "string"
                },
                "tvl": {
                    "type": "string"
                },
                "volume": {
                    "type": "string"
                }
            }
        },
        "dashboard.RecentRes": {
            "type": "object",
            "properties": {
                "fee": {
                    "type": "string"
                },
                "feeChangeRate": {
                    "type": "number"
                },
                "tvl": {
                    "type": "string"
                },
                "tvlChangeRate": {
                    "type": "number"
                },
                "volume": {
                    "type": "string"
                },
                "volumeChangeRate": {
                    "type": "number"
                }
            }
        },
        "dashboard.StatisticRes": {
            "type": "object",
            "properties": {
                "addressCounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.AddressCountRes"
                    }
                },
                "fees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.FeeRes"
                    }
                },
                "txCounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dashboard.TxCountRes"
                    }
                }
            }
        },
        "dashboard.TVLRes": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "tvl": {
                    "type": "string"
                }
            }
        },
        "dashboard.TokenRes": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "priceChange": {
                    "type": "number"
                },
                "tvl": {
                    "type": "string"
                },
                "volume": {
                    "type": "string"
                }
            }
        },
        "dashboard.TxCountRes": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "txCount": {
                    "type": "integer"
                }
            }
        },
        "dashboard.TxRes": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "asset0amount": {
                    "type": "string"
                },
                "asset1amount": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "totalValue": {
                    "type": "string"
                }
            }
        },
        "dashboard.VolumeRes": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "volume": {
                    "type": "string"
                }
            }
        },
...
        "dezswap.AssetInfoTokenRes": {
            "type": "object",
            "properties": {
                "native_token": {
                    "$ref": "#/definitions/dezswap.NativeTokenAssetInfoRes"
                },
                "token": {
                    "$ref": "#/definitions/dezswap.TokenAssetInfoRes"
                }
            }
        },
        "dezswap.NativeTokenAssetInfoRes": {
            "type": "object",
            "properties": {
                "denom": {
                    "type": "string"
                }
            }
        },
        "dezswap.TokenAssetInfoRes": {
            "type": "object",
            "properties": {
                "contract_addr": {
                    "type": "string"
                }
            }
        },
        "httputil.BadRequestError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "bad request"
                }
            }
        },
        "httputil.InternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "httputil.NotFoundError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "message": {
                    "type": "string",
                    "example": "not found"
                }
            }
        }
    }
}